--- packages/java/engine-core/src/test/resources/META-INF/services/com.vaadin.hilla.engine.CustomEngineConfiguration
@@ -1,1 +1,0 @@
-com.vaadin.hilla.engine.EngineConfigurationTest$FirstLoadedConfiguration
