+++ packages/java/engine-core/src/test/resources/META-INF/services/com.vaadin.hilla.engine.EngineConfiguration
@@ -1,0 +1,1 @@
+com.vaadin.hilla.engine.EngineAutoConfigurationTest$FirstLoadedConfiguration
